@@ -8,7 +8,7 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8.0.2",
+        "php": "^8.1",
         "babenkoivan/elastic-migrations": "^1.3",
         "babenkoivan/elastic-scout-driver": "^1.2",
         "babenkoivan/elastic-scout-driver-plus": "^2.3",
@@ -16,18 +16,17 @@
         "backpack/filemanager": "^2.0",
         "backpack/logmanager": "^4.0",
         "backpack/pro": "1.2.8",
-        "doctrine/dbal": "^2.5",
-        "fruitcake/laravel-cors": "^2.0",
+        "doctrine/dbal": "^3.5",
         "guzzlehttp/guzzle": "^7.2",
-        "laravel/framework": "^9.0",
-        "laravel/sanctum": "^3.0",
+        "laravel/framework": "^10.10",
+        "laravel/sanctum": "^3.3",
         "laravel/scout": "^9.0",
         "laravel/telescope": "^4.3",
-        "laravel/tinker": "^2.7",
+        "laravel/tinker": "^2.8",
         "league/flysystem-sftp": "^3.0",
         "mcamara/laravel-localization": "^1.6",
         "sentry/sentry-laravel": "^3.3",
-        "spatie/laravel-medialibrary": "^9.4",
+        "spatie/laravel-medialibrary": "^11.0",
         "spatie/laravel-missing-page-redirector": "^2.7",
         "spatie/laravel-sluggable": "^3.5",
         "spatie/laravel-tags": "^4.5",
@@ -40,17 +39,18 @@
         "deployer/deployer": "^7.3",
         "fakerphp/faker": "^1.9.1",
         "laravel/pint": "^1.0",
-        "laravel/sail": "^1.0.1",
+        "laravel/sail": "^1.18",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^6.1",
-        "phpunit/phpunit": "^9.5.10",
-        "spatie/laravel-ignition": "^1.0"
+        "nunomaduro/collision": "^7.0",
+        "phpunit/phpunit": "^10.1",
+        "spatie/laravel-ignition": "^2.0"
     },
     "config": {
         "optimize-autoloader": true,
         "preferred-install": "dist",
         "sort-packages": true,
         "allow-plugins": {
+            "pestphp/pest-plugin": true,
             "php-http/discovery": true
         }
     },
@@ -71,7 +71,7 @@
             "Tests\\": "tests/"
         }
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true,
     "scripts": {
         "post-install-cmd": [
